Added manifest raw asset entry parsing
