+++ dlasset/audio/main.py
@@ -1,0 +1,6 @@
+"""Implementations to extract audio."""
+__all__ = ("extract_audio",)
+
+
+def extract_audio() -> None:
+    """Extract audio."""
