+++ dlasset/manifest/model/__init__.py
@@ -1,0 +1,4 @@
+"""Manifest file model class."""
+from .entry import ManifestEntry, ManifestRawEntry
+from .locale import ManifestLocale
+from .main import Manifest
